Created the Entity Config
Created the Entities folder. Created the EntityConfig header file in the folder. Created the Entity namespace and created the EntityType enum class (PLAYER, ENEMY, BUNKER, BULLETS, POWERUPS) in the namespace in the header file.

Made changes to the BulletController, BulletModel and the different Bullet Type controller header and source files by adding EntityType object as parameter in the constructors to specify owner entity, by declaring and defining functions to get the owner entity in the BulletModel and BulletController header and source files.

Made changes to the BulletService header and source files by adding EntityType object as parameter in the methods to create and spawn bullet.