Created the ICollider interface and the Collision Service
The ICollider interface is used to provide a base template which can be used to check collisions between all the entities in the game that are able to collide.

The Collision Service is responsible for checking whether two different entities in the game have collided or not and then to execute certain functions accordingly.

Created the Collisions folder. Created the ICollider header and source files in the Collisions folder. Created the Collision namespace and created the ICollider interface in the namespace in and declared necessary pure virtual functions in the header file to return the sprite of the colliding entity and perform certain actions when two entities have collided which will be implemented by classes implementing the interface. Created enum class for collisions states (ENABLED, DISABLED) and declared and defined necessary functions to enable, disable and get the state of the collision in the header and source files.

Created the CollisionService header and source files in the Collisions folder. Created the CollisionService class and forward declared the ICollider interface in the Collision namespace in the header file. Declared and defined vector of ICollider pointers and necessary functions to check for collisions between two entities and process them accordingly in the header and source files.