Set up Collisions for the Player
Made changes to the PlayerController header file by defining the PlayerController class as a class that implements the ICollider interface which will implement the two pure virtual functions of the interface. Declared necessary variables and functions to fire bullets and process Powerup types (shield, triple laser, rapid fire), Bullet and Enemy collisions with the player.

Made changes to the PlayerModel header file by adding another state in the Player State enum class (FROZEN) and by declaring necessary variables and functions which will be called by the PlayerController to set powerup durations and process collisions.

Declared methods to enable the powerups for the Player and to reset them in the PlayerService header file.

Declared a function to get the Player sprite in the PlayerView header file.