Implemented Collisions for bullets
Made changes to the BulletController header header file by defining BulletController class as a class which implements the ICollider interface. Declared and defined the interface's pure virtual functions and the necessary functions to process Bullet collisions with bullets, enemies, players and bunkers in the header and source files.

Declared and defined vector of pointers of IProjectile interface to instances of BulletController class to store the bullets to be removed and the necessary functions to check whether bullet is valid bullet and to destroy the BulletController instances stored in the vector in the BulletService header and source files.

Declared and defined function to get the bullet sprite in the BulletView header and source files

Removed the bullet type texture path variable definitions from the Config file. Created BulletConfig class and declared the bullet type texture path variables and function to get the bullet type texture path in the BulletConfig header file. Created the Bullet Config source file and defined the variables and functions created in the header file.