Implemented Collisions for Enemies and Powerups
Made changes to the EnemyController header file by defining EnemyController class as a class that implements the ICollider interface. Declared and defined the interface's pure virtual functions in a way which destroys enemy on collision with a bullet or player and a function to destroy enemy in the header and source files.

Declared and defined vector to store the Enemy type Controller instances to be deleted and function to delete the instances in the vector in the EnemyService header and source files.

Declared and defined function to get the Enemy sprite in the EnemyView header and source files.

Defined the function to spawn powerup on collision with bullet in the UFOController source file.

Made changes to the PowerupController header file by defining the PowerupController class as a class that implements the ICollider interface. Declared a pure virtual function to apply power up to the player and declared and defined the interface's pure virtual functions in a way where the function to apply power up will be called on collision with the player in the header and source files.

Declared and defined vector to store instances of PowerupController class to be deleted and function to delete the instances in the vector in the PowerupService header and source files.

Declared and defined function to get the Powerup sprite in the PowerupView header and source files.

Declared and defined the function to apply the appropriate powerup in the Powerup type Controllers header and source files.