Implemented the Gameplay UI
Created GameplayUI folder in the UI folder. Created the GameplayUIController header and source files in the folder. Created the GameplayUI namespace and created the GameplayUI class which implements the IUIController interface in the namepsace in the header file. Declared and defined pointers to instances of the ImageView and TextView classes and necessary variables and functions to update and display enemies killed in the form of text and display player lives in the form of images in the game window in the header and source files.

Declared PlayerController class as friend class, two GameplayUIController functions to update enemies killed and display player lives as friend functions and necessary variables to store maximum player lives, player lives left and enemies killed in the PlayerModel header file. Initialized the player lives and enemies killed variables in the PlayerModel source file.

Declared and defined the functions to decrease player lives which will be called when player collides with laser bullet, torpedo or enemy and to increase number of enemies killed in the PlayerController header and source files.

Declared and defined function to increase enemies killed which calls the PlayerController function to increase enemies killed in the PlayerService header and source files.

Added a call to the PlayerService function to increase enemies killed when enemy is destroyed in the EnemyController source file.

Linked the GameplayUIController to the UIService by including the GameplayUIController header file in the UIService header file and declaring and defining pointer to instance of the GameplayUIController class and using the pointer to call GameplayUIController functions in the UIService header and source files.