Created the Splash Screen
Created a Splash Screen folder. Created the SplashScreenUIController header and source files in the folder. Created the SplashScreen namespace and created the SplashScreenUIController class in the namespace as a class that implements the IUIController interface in the header file. Declared and defined the necessary variables and functions to display the Game's intro image for a specified amount of time and then switch to main menu in the header and source files.

Linked the SplashScreenUI Controller to the UI Service by including the SplashScreenUIController header file in the UI Service header file.

Made changes to the UIService header and source files by declaring and defining the pointer to instance of SplashScreenUIController class and using the pointer to call the relevant functions of the SplashScreenUIController class.

Made changes to the GameService header and source files by removing the method to show Main Menu.