Implemented Image Animation for Splash Screen
Created the AnimatedImageView header and source files. Created the AnimatedImageView class as a class that inherits the ImageView class. Declared and defined necessary functions in the header and source files to register call back functions and to allow the fading in and fading out of the image over a specified period of time using the Clock class provided by SFML library.