Created the Animation Data System
Created the Animation System folder. Created the AnimationSystemConfig and AnimationSystemConfigData header files in the folder.

Created the Animation namespace and created the AnimationSystemConfig class in the namespace in the AnimationSystemConfig header file. Declared necessary variables and functions to define properties of a sprite sheet so it can be used to implement animations.

Created instance of AnimationSystemConfig class in the AnimationSystemConfigData header file which sends the properties of explosion sprite sheet as parameters.