Created the Animation System
Created the AnimationSystem header and source files in the Animations folder. Declared and defined pointer to ImageView class and necessary variables and functions in the AnimationSystem header and source files to display animation over a specified amount of time using the sprite sheet defined in the AnimationSystemConfig header file.

Declared and defined functions to set texture of each tile or part of the sprite sheet and to set the dimensions of the entire sprite sheet and of each part of the sheet/frame in the ImageView header and source files.

Personal changes:
Made changes in the SoundService header and source files and called function to implement sounds of firing bullets in the PlayerController source file when the player ship fires.