Created the Animation Service and linked it to Service Locator
Created the AnimationService header and source files in the Animations folder. Declared and defined enum class for animation types like EXPLOSION, two vectors of pointers to instances of AnimationSystem class for storing the animations and the animation instances to be deleted and necessary functions to spawn and destroy the animations in the header and source files.

Linked the Animation Service to the Service Locator by including the AnimationService header file in the ServiceLocator header file and declaring and defining pointer to instance of the AnimationService class to call appropriate functions of the Animation Service and a function to retrieve the AnimationService instance in the ServiceLocator header and source files.