Improvements
Declared and defined method to get Animation texture path in animation Service and send it to Animation System constructor. Had to do this because there was an error where the game was crashing when trying to access the texture path variable of AnimationSystemConfig class using the AnimationSystemConfig object in the AnimationSystem file.

Made changes to enemy collision in EnemyController so that any enemy get destroyed on collision with bunker but bunker only gets destroyed when collided with SUBZERO enemy.